Refactor: Rename arbitrator package to disputeresolvers
