+++ desktop/src/main/java/bisq/desktop/main/disputes/disputeresolvers/ArbitratorDisputeView.fxml
@@ -19,7 +19,7 @@
 
 <?import javafx.scene.layout.VBox?>
 <?import javafx.geometry.Insets?>
-<VBox fx:id="root" fx:controller="bisq.desktop.main.disputes.arbitrator.ArbitratorDisputeView"
+<VBox fx:id="root" fx:controller="bisq.desktop.main.disputes.disputeresolvers.ArbitratorDisputeView"
       spacing="10"
       xmlns:fx="http://javafx.com/fxml">
     <padding>
